working on redeem coupon

+++ lib/features/auth/domain/entities/user_detail.dart
@@ -65,7 +65,7 @@ class UserDetail extends Equatable {
   final String? uuid;
   final String? avatar;
   final double? balance;
-  final int? point;
+  final double? point;
   final String? dob;
   final String? gender;
   final String? nationality;
